requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
accelerate==0.25.0
gunicorn==21.2.0